ES6 modulize core/data
