+++ test/qunit/unit-es5/test-class.js
@@ -444,7 +444,7 @@ test('_isCCClass', function () {
     function ctor () {
         this.foo = 0;
     }
-    cc.Class._fastDefine('T', ctor, ['foo']);
+    cc.Class.fastDefine('T', ctor, ['foo']);
     strictEqual(cc.Class._isCCClass(ctor), false, 'fastDefined ctor should not recognized as CCClass');
 
     cc.js.unregisterClass(ctor);
